Relocate COSMOS plugin generator classes to respective components
Clean up COSMOS plugin generator inheritance

Correct errors picked up by style check

Correct errors picked up by style check

Correct additional errors picked up by style check
