+++ gen/templates/assembly/name_ccsds_cosmos_plugin.txt
