+++ src/components/ccsds_packetizer/gen/templates/assembly/name_ccsds_cosmos_telemetry.txt
